[BE] FIX: 피드백 반영 #1038
