+++ backend/src/main/java/org/ftclub/cabinet/user/service/UserFacadeService.java
@@ -1,6 +1,7 @@
 package org.ftclub.cabinet.user.service;
 
 import java.util.Date;
+import java.util.List;
 import org.ftclub.cabinet.cabinet.domain.LentType;
 import org.ftclub.cabinet.dto.BlockedUserPaginationDto;
 import org.ftclub.cabinet.dto.MyCabinetInfoResponseDto;
@@ -9,6 +10,7 @@ import org.ftclub.cabinet.dto.UserCabinetPaginationDto;
 import org.ftclub.cabinet.dto.UserProfilePaginationDto;
 import org.ftclub.cabinet.dto.UserSessionDto;
 import org.ftclub.cabinet.user.domain.AdminRole;
+import org.ftclub.cabinet.user.domain.User;
 import org.ftclub.cabinet.user.domain.UserRole;
 
 public interface UserFacadeService {
@@ -62,6 +64,13 @@ public interface UserFacadeService {
 	 * @return {@link MyCabinetInfoResponseDto} 본인의 대여 정보와 캐비넷 정보
 	 */
 	MyCabinetInfoResponseDto getMyLentAndCabinetInfo(Long userId);
+
+	/**
+	 * 모든 유저의 정보를 가져옵니다.
+	 *
+	 * @return 모든 유저의 정보를 가져옵니다.
+	 */
+	List<User> getAllUsers();
 
 	/**
 	 * 유저가 존재하는지 확인합니다.
@@ -144,5 +153,5 @@ public interface UserFacadeService {
 	 * @param userId 유저 고유 아이디
 	 * @param today  현재 날짜
 	 */
-	void unbanUser(Long userId, Date today);
+	void deleteRecentBanHistory(Long userId, Date today);
 }
