+++ backend/src/main/java/org/ftclub/cabinet/user/service/UserService.java
@@ -1,10 +1,8 @@
 package org.ftclub.cabinet.user.service;
 
 import java.util.Date;
-import java.util.List;
 import org.ftclub.cabinet.cabinet.domain.LentType;
 import org.ftclub.cabinet.user.domain.AdminRole;
-import org.ftclub.cabinet.user.domain.User;
 import org.ftclub.cabinet.user.domain.UserRole;
 
 public interface UserService {
@@ -28,7 +26,7 @@ public interface UserService {
 
 	void banUser(Long userId, LentType lentType, Date startedAt, Date endedAt, Date expiredAt);
 
-	void unbanUser(Long userId, Date today);
+	void deleteRecentBanHistory(Long userId, Date today);
 
 	String getUserEmail(Long userId);
 
@@ -48,11 +46,4 @@ public interface UserService {
 	 * @return 정지 상태인 경우 true, 정지 상태가 아닌 경우 false
 	 */
 	boolean checkUserIsBanned(Long userId, Date today);
-
-	/**
-	 * 모든 유저의 정보를 가져옵니다.
-	 *
-	 * @return 모든 유저의 정보를 가져옵니다.
-	 */
-	List<User> getAllUsers();
 }
